New function for generating a projector mask
Mask function takes an array of eccentricities at each 15* angle (See
func for the specific angle order). Generates a stencil mask that can
then be used by other presentation programs. For example if you do:
mglProjStencil(eccArray);
mglClearScreen(0);
mglStencilSelect(1);
mglFillRect(0,0,[15 15],[.5 .5 .5]);
mglStencilSelect(0);
mglFlush
It’s equivalent to clearing the masked part of the screen, and setting
the rest of the screen to black so that the projector won’t illuminate
it.