fix bug with mean calculation in fisher linear discriminant
When a class only has two instances and is used in leave-one-out. Only
one instance (x1, 1 by N dimensional vector) is used for training
fisher linear discriminant. Fisher linear discriminant needs the mean
instance coordinates to calculate the classifier weights. The bug is
that mean(x1), the average coordinates (a scalar) is wrong in that
case, but mean(x1,1) the average instance coordinates is correct.